sociology_kb_starter/requirements.txt
streamlit>=1.44.0
pydantic>=2.6.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
pandas>=2.2.0
PyMuPDF>=1.24.0
requests>=2.31.0
openai>=1.30.0
anthropic>=0.25.0
matplotlib>=3.8.0
watchdog>=4.0.0
numpy>=1.26.0

pytest>=8.2.0

streamlit-agraph>=0.0.45
mcp>=1.2.0
